fix: ensure all loaded doubles are actually converted floats

+++ src/main/java/me/contaria/standardsettings/StandardSettingsConfig.java
@@ -193,7 +193,8 @@ public class StandardSettingsConfig implements SpeedrunConfig {
                     }
                     double value;
                     try {
-                        value = Double.parseDouble(string);
+                        // read only up to float precision, matching the behavior of GameOptions#load
+                        value = Float.parseFloat(string);
                     } catch (NumberFormatException e) {
                         return;
                     }
